refactor: 깔끔한 아키텍처 준수를 위한 패키지 및 클래스 네이밍 재구성
