+++ src/main/java/org/c4marathon/assignment/domain/repository/SavingAccountRepository.java
@@ -1,0 +1,7 @@
+package org.c4marathon.assignment.domain.repository;
+
+import org.c4marathon.assignment.domain.model.account.SavingAccount;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SavingAccountRepository extends JpaRepository<SavingAccount, Long> {
+}
