+++ src/main/java/org/c4marathon/assignment/domain/repository/mainaccount/MainAccountQueryRepository.java
@@ -1,0 +1,5 @@
+package org.c4marathon.assignment.domain.repository.mainaccount;
+
+public interface MainAccountQueryRepository {
+	int conditionalFastCharge(Long id, Long amount, Long minRequiredBalance, Long limit);
+}
