+++ src/main/java/org/c4marathon/assignment/dto/account/AccountDto.java
@@ -1,6 +1,6 @@
 package org.c4marathon.assignment.dto.account;
 
-import org.c4marathon.assignment.domain.account.Account;
+import org.c4marathon.assignment.domain.model.account.Account;
 
 public record AccountDto(Long id, Long memberId, Long balance) {
     public AccountDto(Account account) {
