--- src/main/java/org/c4marathon/assignment/repository/SavingAccountRepository.java
@@ -1,7 +1,0 @@
-package org.c4marathon.assignment.repository;
-
-import org.c4marathon.assignment.domain.account.SavingAccount;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SavingAccountRepository extends JpaRepository<SavingAccount, Long> {
-}
