--- src/main/java/org/c4marathon/assignment/repository/mainAccount/MainAccountQueryRepository.java
@@ -1,5 +1,0 @@
-package org.c4marathon.assignment.repository.mainAccount;
-
-public interface MainAccountQueryRepository {
-	int conditionalFastCharge(Long id, Long amount, Long minRequiredBalance, Long limit);
-}
