+++ src/main/java/org/c4marathon/assignment/scheduler/AccountScheduler.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.scheduler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.c4marathon.assignment.repository.mainAccount.MainAccountRepository;
+import org.c4marathon.assignment.domain.repository.mainaccount.MainAccountRepository;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
